Move the templates application function into its own module
