+++ lib/pltr/v2/helpers/__tests__/beats.test.js
@@ -1,6 +1,3 @@
-import { omit } from 'lodash'
-
-import { tree } from '../../index'
 import { children, findNode, newTree, nodeParent } from '../../reducers/tree'
 import { beat } from '../../store/initialState'
 import {
@@ -9,29 +6,7 @@ import {
   adjustHierarchyLevels,
   maxDepth,
   numberOfPriorChildrenAtSameDepth,
-  applyTemplate,
-  beatsLength,
 } from '../beats'
-import {
-  // Templates
-  one_level_plotline_template,
-  two_level_plotline_template,
-  three_level_plotline_template,
-  one_level_biased_top_plotline_template,
-  two_level_biased_top_plotline_template,
-  three_level_biased_top_plotline_template,
-  one_level_biased_middle_plotline_template,
-  two_level_biased_middle_plotline_template,
-  three_level_biased_middle_plotline_template,
-  one_level_biased_bottom_plotline_template,
-  two_level_biased_bottom_plotline_template,
-  three_level_biased_bottom_plotline_template,
-
-  // Files
-  file_with_one_level,
-  file_with_two_levels,
-  file_with_three_levels,
-} from './fixtures'
 
 describe('addLevelToHierarchy', () => {
   describe('given the empty tree', () => {
@@ -433,1852 +408,6 @@ describe('numberOfPriorChildrenAtSameDepth', () => {
         expect(numberOfPriorChildrenAtSameDepth(tree, sortedBeats, 10)).toEqual(3)
         expect(numberOfPriorChildrenAtSameDepth(tree, sortedBeats, 11)).toEqual(4)
         expect(numberOfPriorChildrenAtSameDepth(tree, sortedBeats, 12)).toEqual(3)
-      })
-    })
-  })
-})
-
-const withoutLineBeatOrCardIds = (x) => omit(x, 'id', 'lineId', 'beatId')
-
-// Consider also testing:
-//  - the names of inserted beats,
-//  - the exact positions that cards end up in,
-//  - the order of inserted plot lines, and
-//  - edge cases.
-describe('applyTemplate', () => {
-  describe('given a file with one level', () => {
-    describe('and a template with one level', () => {
-      describe('and the template is biased to the "top"', () => {
-        const result = applyTemplate(file_with_one_level, 1, one_level_biased_top_plotline_template)
-        it('should have more cards', () => {
-          expect(result.cards.length).toBeGreaterThanOrEqual(file_with_one_level.cards.length)
-        })
-        it('should have the cards it used to have', () => {
-          expect(result.cards).toEqual(expect.arrayContaining(file_with_one_level.cards))
-        })
-        it('should have the cards in the template', () => {
-          expect(result.cards.map(withoutLineBeatOrCardIds)).toEqual(
-            expect.arrayContaining(
-              one_level_biased_top_plotline_template.templateData.cards.map(
-                withoutLineBeatOrCardIds
-              )
-            )
-          )
-        })
-        it('should have at least as many beats as the template', () => {
-          const templateNumberOfBeats = beatsLength(
-            one_level_biased_top_plotline_template.templateData.beats['1']
-          )
-          const resultNumberOfBeats = beatsLength(result.beats['1'])
-          expect(resultNumberOfBeats).toBeGreaterThanOrEqual(templateNumberOfBeats)
-        })
-        it('should add the plotlines from the template', () => {
-          expect(result.lines).toEqual(
-            expect.arrayContaining(one_level_biased_top_plotline_template.templateData.lines)
-          )
-        })
-        it('should have the plotlines that were there originally', () => {
-          expect(result.lines).toEqual(expect.arrayContaining(file_with_one_level.lines))
-        })
-        it('should insert cards on the plotlines from the template', () => {
-          const templateLines = result.lines
-            .map((line) => {
-              return line.id
-            })
-            .filter((lineId) => {
-              return (
-                file_with_one_level.lines
-                  .map((line) => {
-                    return line.id
-                  })
-                  .indexOf(lineId) === -1
-              )
-            })
-          const addedCards = result.cards.filter((card) => {
-            return templateLines.indexOf(card.lineId) !== -1
-          })
-          expect(one_level_biased_top_plotline_template.templateData.cards.length).toEqual(
-            addedCards.length
-          )
-          expect(addedCards.map(withoutLineBeatOrCardIds)).toEqual(
-            expect.arrayContaining(
-              one_level_biased_top_plotline_template.templateData.cards.map(
-                withoutLineBeatOrCardIds
-              )
-            )
-          )
-        })
-      })
-      describe('and the template is biased to the "middle"', () => {
-        const result = applyTemplate(
-          file_with_one_level,
-          1,
-          one_level_biased_middle_plotline_template
-        )
-        it('should have more cards', () => {
-          expect(result.cards.length).toBeGreaterThanOrEqual(file_with_one_level.cards.length)
-        })
-        it('should have the cards it used to have', () => {
-          expect(result.cards).toEqual(expect.arrayContaining(file_with_one_level.cards))
-        })
-        it('should have the cards in the template', () => {
-          expect(result.cards.map(withoutLineBeatOrCardIds)).toEqual(
-            expect.arrayContaining(
-              one_level_biased_middle_plotline_template.templateData.cards.map(
-                withoutLineBeatOrCardIds
-              )
-            )
-          )
-        })
-        it('should have at least as many beats as the template', () => {
-          const templateNumberOfBeats = beatsLength(
-            one_level_biased_middle_plotline_template.templateData.beats['1']
-          )
-          const resultNumberOfBeats = beatsLength(result.beats['1'])
-          expect(resultNumberOfBeats).toBeGreaterThanOrEqual(templateNumberOfBeats)
-        })
-        it('should add the plotlines from the template', () => {
-          expect(result.lines).toEqual(
-            expect.arrayContaining(one_level_biased_middle_plotline_template.templateData.lines)
-          )
-        })
-        it('should have the plotlines that were there originally', () => {
-          expect(result.lines).toEqual(expect.arrayContaining(file_with_one_level.lines))
-        })
-        it('should insert cards on the plotlines from the template', () => {
-          const templateLines = result.lines
-            .map((line) => {
-              return line.id
-            })
-            .filter((lineId) => {
-              return (
-                file_with_one_level.lines
-                  .map((line) => {
-                    return line.id
-                  })
-                  .indexOf(lineId) === -1
-              )
-            })
-          const addedCards = result.cards.filter((card) => {
-            return templateLines.indexOf(card.lineId) !== -1
-          })
-          expect(one_level_biased_middle_plotline_template.templateData.cards.length).toEqual(
-            addedCards.length
-          )
-          expect(addedCards.map(withoutLineBeatOrCardIds)).toEqual(
-            expect.arrayContaining(
-              one_level_biased_middle_plotline_template.templateData.cards.map(
-                withoutLineBeatOrCardIds
-              )
-            )
-          )
-        })
-      })
-      describe('and the template is biased to the "bottom"', () => {
-        const result = applyTemplate(
-          file_with_one_level,
-          1,
-          one_level_biased_bottom_plotline_template
-        )
-        it('should have more cards', () => {
-          expect(result.cards.length).toBeGreaterThanOrEqual(file_with_one_level.cards.length)
-        })
-        it('should have the cards it used to have', () => {
-          expect(result.cards).toEqual(expect.arrayContaining(file_with_one_level.cards))
-        })
-        it('should have the cards in the template', () => {
-          expect(result.cards.map(withoutLineBeatOrCardIds)).toEqual(
-            expect.arrayContaining(
-              one_level_biased_bottom_plotline_template.templateData.cards.map(
-                withoutLineBeatOrCardIds
-              )
-            )
-          )
-        })
-        it('should have at least as many beats as the template', () => {
-          const templateNumberOfBeats = beatsLength(
-            one_level_biased_bottom_plotline_template.templateData.beats['1']
-          )
-          const resultNumberOfBeats = beatsLength(result.beats['1'])
-          expect(resultNumberOfBeats).toBeGreaterThanOrEqual(templateNumberOfBeats)
-        })
-        it('should add the plotlines from the template', () => {
-          expect(result.lines).toEqual(
-            expect.arrayContaining(one_level_biased_bottom_plotline_template.templateData.lines)
-          )
-        })
-        it('should have the plotlines that were there originally', () => {
-          expect(result.lines).toEqual(expect.arrayContaining(file_with_one_level.lines))
-        })
-        it('should insert cards on the plotlines from the template', () => {
-          const templateLines = result.lines
-            .map((line) => {
-              return line.id
-            })
-            .filter((lineId) => {
-              return (
-                file_with_one_level.lines
-                  .map((line) => {
-                    return line.id
-                  })
-                  .indexOf(lineId) === -1
-              )
-            })
-          const addedCards = result.cards.filter((card) => {
-            return templateLines.indexOf(card.lineId) !== -1
-          })
-          expect(one_level_biased_bottom_plotline_template.templateData.cards.length).toEqual(
-            addedCards.length
-          )
-          expect(addedCards.map(withoutLineBeatOrCardIds)).toEqual(
-            expect.arrayContaining(
-              one_level_biased_bottom_plotline_template.templateData.cards.map(
-                withoutLineBeatOrCardIds
-              )
-            )
-          )
-        })
-      })
-    })
-    describe('and a template with two levels', () => {
-      describe('and the template is biased to the "top"', () => {
-        const result = applyTemplate(file_with_one_level, 1, two_level_biased_top_plotline_template)
-        const topLevelBeats = tree.children(
-          two_level_biased_top_plotline_template.templateData.beats['1'],
-          null
-        )
-        it('should have more cards', () => {
-          expect(result.cards.length).toBeGreaterThanOrEqual(file_with_one_level.cards.length)
-        })
-        it('should have the cards it used to have', () => {
-          expect(result.cards).toEqual(expect.arrayContaining(file_with_one_level.cards))
-        })
-        it('should have the cards in the template', () => {
-          const topLevelBeatIds = topLevelBeats.map(({ id }) => id)
-          const cardsInTemplate = two_level_biased_top_plotline_template.templateData.cards.filter(
-            (card) => {
-              return topLevelBeatIds.indexOf(card.beatId) !== -1
-            }
-          )
-          expect(result.cards.map(withoutLineBeatOrCardIds)).toEqual(
-            expect.arrayContaining(cardsInTemplate.map(withoutLineBeatOrCardIds))
-          )
-        })
-        it('should have at least as many beats as the top level of the template', () => {
-          const templateNumberOfBeats = topLevelBeats.length
-          const resultNumberOfBeats = beatsLength(result.beats['1'])
-          expect(resultNumberOfBeats).toBeGreaterThanOrEqual(templateNumberOfBeats)
-        })
-        it('should add the plotlines from the template', () => {
-          expect(result.lines).toEqual(
-            expect.arrayContaining(two_level_biased_top_plotline_template.templateData.lines)
-          )
-        })
-        it('should have the plotlines that were there originally', () => {
-          expect(result.lines).toEqual(expect.arrayContaining(file_with_one_level.lines))
-        })
-        it('should insert cards on the plotlines from the template', () => {
-          const templateLines = result.lines
-            .map((line) => {
-              return line.id
-            })
-            .filter((lineId) => {
-              return (
-                file_with_one_level.lines
-                  .map((line) => {
-                    return line.id
-                  })
-                  .indexOf(lineId) === -1
-              )
-            })
-          const addedCards = result.cards.filter((card) => {
-            return templateLines.indexOf(card.lineId) !== -1
-          })
-          const topLevelBeatIds = topLevelBeats.map(({ id }) => id)
-          const templateCards = two_level_biased_top_plotline_template.templateData.cards.filter(
-            ({ beatId }) => {
-              return topLevelBeatIds.indexOf(beatId) !== -1
-            }
-          )
-          expect(two_level_biased_top_plotline_template.templateData.cards.length).toEqual(
-            addedCards.length
-          )
-          expect(addedCards.map(withoutLineBeatOrCardIds)).toEqual(
-            expect.arrayContaining(templateCards.map(withoutLineBeatOrCardIds))
-          )
-        })
-      })
-      describe('and the template is biased to the middle', () => {
-        const result = applyTemplate(
-          file_with_one_level,
-          1,
-          two_level_biased_middle_plotline_template
-        )
-        const bottomLevelBeats = tree.children(
-          two_level_biased_middle_plotline_template.templateData.beats['1'],
-          null
-        )
-        it('should have more cards', () => {
-          expect(result.cards.length).toBeGreaterThanOrEqual(file_with_one_level.cards.length)
-        })
-        it('should have the cards it used to have', () => {
-          expect(result.cards).toEqual(expect.arrayContaining(file_with_one_level.cards))
-        })
-        it('should have the cards in the template from the bottom level', () => {
-          const bottomLevelBeatIds = bottomLevelBeats.map(({ id }) => id)
-          const cardsInBookFromTemplate = result.cards.filter((card) => {
-            return bottomLevelBeatIds.indexOf(card.beatId) !== -1
-          })
-          expect(result.cards).toEqual(expect.arrayContaining(cardsInBookFromTemplate))
-        })
-        it('should have at least as many beats as the bottom level of the template', () => {
-          const templateNumberOfBeats = bottomLevelBeats.length
-          const resultNumberOfBeats = beatsLength(result.beats['1'])
-          expect(resultNumberOfBeats).toBeGreaterThanOrEqual(templateNumberOfBeats)
-        })
-        it('should add the plotlines from the template', () => {
-          expect(result.lines).toEqual(
-            expect.arrayContaining(two_level_biased_middle_plotline_template.templateData.lines)
-          )
-        })
-        it('should have the plotlines that were there originally', () => {
-          expect(result.lines).toEqual(expect.arrayContaining(file_with_one_level.lines))
-        })
-        it('should insert cards on the plotlines from the template', () => {
-          const templateLines = result.lines
-            .map((line) => {
-              return line.id
-            })
-            .filter((lineId) => {
-              return (
-                file_with_one_level.lines
-                  .map((line) => {
-                    return line.id
-                  })
-                  .indexOf(lineId) === -1
-              )
-            })
-          const addedCards = result.cards.filter((card) => {
-            return templateLines.indexOf(card.lineId) !== -1
-          })
-          const bottomLevelBeatIds = bottomLevelBeats.map(({ id }) => id)
-          const templateCards = two_level_biased_middle_plotline_template.templateData.cards.filter(
-            ({ beatId }) => {
-              return bottomLevelBeatIds.indexOf(beatId) !== -1
-            }
-          )
-          expect(two_level_biased_middle_plotline_template.templateData.cards.length).toEqual(
-            addedCards.length
-          )
-          expect(addedCards.map(withoutLineBeatOrCardIds)).toEqual(
-            expect.arrayContaining(templateCards.map(withoutLineBeatOrCardIds))
-          )
-        })
-      })
-      describe('and the template is biased to the bottom', () => {
-        const result = applyTemplate(
-          file_with_one_level,
-          1,
-          two_level_biased_bottom_plotline_template
-        )
-        const bottomLevelBeats = tree.children(
-          two_level_biased_bottom_plotline_template.templateData.beats['1'],
-          null
-        )
-        it('should have more cards', () => {
-          expect(result.cards.length).toBeGreaterThanOrEqual(file_with_one_level.cards.length)
-        })
-        it('should have the cards it used to have', () => {
-          expect(result.cards).toEqual(expect.arrayContaining(file_with_one_level.cards))
-        })
-        it('should have the cards in the template from the bottom level', () => {
-          const bottomLevelBeatIds = bottomLevelBeats.map(({ id }) => id)
-          const cardsInBookFromTemplate = result.cards.filter((card) => {
-            return bottomLevelBeatIds.indexOf(card.beatId) !== -1
-          })
-          expect(result.cards).toEqual(expect.arrayContaining(cardsInBookFromTemplate))
-        })
-        it('should have at least as many beats as the bottom level of the template', () => {
-          const templateNumberOfBeats = bottomLevelBeats.length
-          const resultNumberOfBeats = beatsLength(result.beats['1'])
-          expect(resultNumberOfBeats).toBeGreaterThanOrEqual(templateNumberOfBeats)
-        })
-        it('should add the plotlines from the template', () => {
-          expect(result.lines).toEqual(
-            expect.arrayContaining(two_level_biased_bottom_plotline_template.templateData.lines)
-          )
-        })
-        it('should have the plotlines that were there originally', () => {
-          expect(result.lines).toEqual(expect.arrayContaining(file_with_one_level.lines))
-        })
-        it('should insert cards on the plotlines from the template', () => {
-          const templateLines = result.lines
-            .map((line) => {
-              return line.id
-            })
-            .filter((lineId) => {
-              return (
-                file_with_one_level.lines
-                  .map((line) => {
-                    return line.id
-                  })
-                  .indexOf(lineId) === -1
-              )
-            })
-          const addedCards = result.cards.filter((card) => {
-            return templateLines.indexOf(card.lineId) !== -1
-          })
-          const bottomLevelBeatIds = bottomLevelBeats.map(({ id }) => id)
-          const templateCards = two_level_biased_bottom_plotline_template.templateData.cards.filter(
-            ({ beatId }) => {
-              return bottomLevelBeatIds.indexOf(beatId) !== -1
-            }
-          )
-          expect(two_level_biased_bottom_plotline_template.templateData.cards.length).toEqual(
-            addedCards.length
-          )
-          expect(addedCards.map(withoutLineBeatOrCardIds)).toEqual(
-            expect.arrayContaining(templateCards.map(withoutLineBeatOrCardIds))
-          )
-        })
-      })
-    })
-    describe('and a template with three levels', () => {
-      describe('and the template is biased to the top', () => {
-        const result = applyTemplate(
-          file_with_one_level,
-          1,
-          three_level_biased_top_plotline_template
-        )
-        const topLevelBeats = tree.children(
-          three_level_biased_top_plotline_template.templateData.beats['1'],
-          null
-        )
-        it('should have more cards', () => {
-          expect(result.cards.length).toBeGreaterThanOrEqual(file_with_one_level.cards.length)
-        })
-        it('should have the cards it used to have', () => {
-          expect(result.cards).toEqual(expect.arrayContaining(file_with_one_level.cards))
-        })
-        it('should have the cards in the template', () => {
-          const topLevelBeatIds = topLevelBeats.map(({ id }) => id)
-          const cardsInBookFromTemplate =
-            three_level_biased_top_plotline_template.templateData.cards.filter((card) => {
-              return topLevelBeatIds.indexOf(card.beatId) !== -1
-            })
-          expect(result.cards.map(withoutLineBeatOrCardIds)).toEqual(
-            expect.arrayContaining(cardsInBookFromTemplate.map(withoutLineBeatOrCardIds))
-          )
-        })
-        it('should have at least as many beats as the top level of the template', () => {
-          const templateNumberOfBeats = topLevelBeats.length
-          const resultNumberOfBeats = beatsLength(result.beats['1'])
-          expect(resultNumberOfBeats).toBeGreaterThanOrEqual(templateNumberOfBeats)
-        })
-        it('should add the plotlines from the template', () => {
-          expect(result.lines).toEqual(
-            expect.arrayContaining(three_level_biased_top_plotline_template.templateData.lines)
-          )
-        })
-        it('should have the plotlines that were there originally', () => {
-          expect(result.lines).toEqual(expect.arrayContaining(file_with_one_level.lines))
-        })
-        it('should insert cards on the plotlines from the template', () => {
-          const templateLines = result.lines
-            .map((line) => {
-              return line.id
-            })
-            .filter((lineId) => {
-              return (
-                file_with_one_level.lines
-                  .map((line) => {
-                    return line.id
-                  })
-                  .indexOf(lineId) === -1
-              )
-            })
-          const addedCards = result.cards.filter((card) => {
-            return templateLines.indexOf(card.lineId) !== -1
-          })
-          const topLevelBeatIds = topLevelBeats.map(({ id }) => id)
-          const templateCards = three_level_biased_top_plotline_template.templateData.cards.filter(
-            ({ beatId }) => {
-              return topLevelBeatIds.indexOf(beatId) !== -1
-            }
-          )
-          expect(three_level_biased_top_plotline_template.templateData.cards.length).toEqual(
-            addedCards.length
-          )
-          expect(addedCards.map(withoutLineBeatOrCardIds)).toEqual(
-            expect.arrayContaining(templateCards.map(withoutLineBeatOrCardIds))
-          )
-        })
-      })
-      describe('and the template is biased to the middle', () => {
-        const result = applyTemplate(
-          file_with_one_level,
-          1,
-          three_level_biased_middle_plotline_template
-        )
-        const topLevelBeats = tree.children(
-          three_level_biased_middle_plotline_template.templateData.beats['1'],
-          null
-        )
-        const middleLevelBeats = topLevelBeats.flatMap((beat) => {
-          return tree.children(result.beats['1'], beat.id)
-        })
-        it('should have more cards', () => {
-          expect(result.cards.length).toBeGreaterThanOrEqual(file_with_one_level.cards.length)
-        })
-        it('should have the cards it used to have', () => {
-          expect(result.cards).toEqual(expect.arrayContaining(file_with_one_level.cards))
-        })
-        it('should have the cards in the template', () => {
-          const middleLevelBeatIds = middleLevelBeats.map(({ id }) => id)
-          const cardsInBookFromTemplate =
-            three_level_biased_middle_plotline_template.templateData.cards.filter((card) => {
-              return middleLevelBeatIds.indexOf(card.beatId) !== -1
-            })
-          expect(result.cards.map(withoutLineBeatOrCardIds)).toEqual(
-            expect.arrayContaining(cardsInBookFromTemplate.map(withoutLineBeatOrCardIds))
-          )
-        })
-        it('should have at least as many beats as the middle level of the template', () => {
-          const templateNumberOfBeats = middleLevelBeats.length
-          const resultNumberOfBeats = beatsLength(result.beats['1'])
-          expect(resultNumberOfBeats).toBeGreaterThanOrEqual(templateNumberOfBeats)
-        })
-        it('should add the plotlines from the template', () => {
-          expect(result.lines).toEqual(
-            expect.arrayContaining(three_level_biased_middle_plotline_template.templateData.lines)
-          )
-        })
-        it('should have the plotlines that were there originally', () => {
-          expect(result.lines).toEqual(expect.arrayContaining(file_with_one_level.lines))
-        })
-        it('should insert cards on the plotlines from the template', () => {
-          const templateLines = result.lines
-            .map((line) => {
-              return line.id
-            })
-            .filter((lineId) => {
-              return (
-                file_with_one_level.lines
-                  .map((line) => {
-                    return line.id
-                  })
-                  .indexOf(lineId) === -1
-              )
-            })
-          const addedCards = result.cards.filter((card) => {
-            return templateLines.indexOf(card.lineId) !== -1
-          })
-          const middleLevelBeatIds = middleLevelBeats.map(({ id }) => id)
-          const templateCards =
-            three_level_biased_middle_plotline_template.templateData.cards.filter(({ beatId }) => {
-              return middleLevelBeatIds.indexOf(beatId) !== -1
-            })
-          expect(three_level_biased_middle_plotline_template.templateData.cards.length).toEqual(
-            addedCards.length
-          )
-          expect(addedCards.map(withoutLineBeatOrCardIds)).toEqual(
-            expect.arrayContaining(templateCards.map(withoutLineBeatOrCardIds))
-          )
-        })
-      })
-      describe('and the template is biased to the bottom', () => {
-        const result = applyTemplate(
-          file_with_one_level,
-          1,
-          three_level_biased_bottom_plotline_template
-        )
-        const topLevelBeats = tree.children(
-          three_level_biased_bottom_plotline_template.templateData.beats['1'],
-          null
-        )
-        const middleLevelBeats = topLevelBeats.flatMap((beat) => {
-          return tree.children(result.beats['1'], beat.id)
-        })
-        const bottomLevelBeats = middleLevelBeats.flatMap((beat) => {
-          return tree.children(result.beats['1'], beat.id)
-        })
-        it('should have more cards', () => {
-          expect(result.cards.length).toBeGreaterThanOrEqual(file_with_one_level.cards)
-        })
-        it('should have the cards it used to have', () => {
-          expect(result.cards).toEqual(expect.arrayContaining(file_with_one_level.cards))
-        })
-        it('should have the cards in the template', () => {
-          const bottomLevelBeatIds = bottomLevelBeats.map(({ id }) => {
-            return id
-          })
-          const cardsInBookFromTemplate = three_level_biased_bottom_plotline_template.cards.filter(
-            ({ beatId }) => {
-              return bottomLevelBeatIds.indexOf(beatId) !== -1
-            }
-          )
-          expect(result.cards.map(withoutLineBeatOrCardIds)).toEqual(
-            expect.arrayContaining(cardsInBookFromTemplate.map(withoutLineBeatOrCardIds))
-          )
-        })
-        it('should have at least as many beats as the bottom level of the template', () => {
-          const templateNumberOfBeats = bottomLevelBeats.length
-          const resultNumberOfBeats = beatsLength(result.beats['1'])
-          expect(resultNumberOfBeats).toBeGreaterThanOrEqual(templateNumberOfBeats)
-        })
-        it('should add the plotlines from the template', () => {
-          expect(result.lines).toEqual(
-            expect.araryContaining(three_level_biased_bottom_plotline_template)
-          )
-        })
-        it('should have the plotlines that were there originally', () => {
-          expect(result.lines).toEqual(expect.arrayContaining(file_with_one_level.lines))
-        })
-        it('should insert cards on teh plotlines from the template', () => {
-          const templateLines = three_level_biased_bottom_plotline_template.templateData.lines.map(
-            (line) => {
-              return line.id
-            }
-          )
-          const addedCards = result.cards.filter((card) => {
-            return templateLines.indexOf(card.lineId) !== -1
-          })
-          expect(three_level_biased_bottom_plotline_template.templateData.cards.length).toEqual(
-            addedCards.length
-          )
-          expect(addedCards.map(withoutLineBeatOrCardIds)).toEqual(
-            expect.arrayContaining(
-              three_level_biased_bottom_plotline_template.templateData.cards.map(
-                withoutLineBeatOrCardIds
-              )
-            )
-          )
-        })
-      })
-    })
-  })
-  describe('given a file with two levels', () => {
-    describe('and a template with one level', () => {
-      describe('and the template is biased to the "top"', () => {
-        const result = applyTemplate(
-          file_with_two_levels,
-          1,
-          one_level_biased_top_plotline_template
-        )
-        const templateBeats = tree.children(
-          one_level_biased_top_plotline_template.templateData.beats['1'],
-          null
-        )
-        it('should have more cards', () => {
-          expect(result.cards.length).toBeGreaterThanOrEqual(file_with_two_levels.cards.length)
-        })
-        it('should have the cards it used to have', () => {
-          expect(result.cards).toEqual(expect.arrayContaining(file_with_two_levels.cards))
-        })
-        it('should have the cards in the template', () => {
-          const templateBeatIds = templateBeats.map(({ beatId }) => {
-            return beatId
-          })
-          const cardsInBookFromTemplate =
-            one_level_biased_top_plotline_template.templateData.cards.filter((card) => {
-              return templateBeatIds.indexOf(card.beatId) !== -1
-            })
-          expect(result.cards.map(withoutLineBeatOrCardIds)).toEqual(
-            expect.arrayContaining(cardsInBookFromTemplate.map(withoutLineBeatOrCardIds))
-          )
-        })
-        it('should have at least as many beats as the top level of the template', () => {
-          const templateNumberOfBeats = templateBeats.length
-          const resultNumberOfBeats = beatsLength(result.beats['1'])
-          expect(resultNumberOfBeats).toBeGreaterThanOrEqual(templateNumberOfBeats)
-        })
-        it('should have the plotlines that were there originally', () => {
-          expect(result.lines).toEqual(expect.arrayContaining(file_with_two_levels.lines))
-        })
-        it('should insert cards on the plotlines from the template', () => {
-          const templateLines = result.lines
-            .map((line) => {
-              return line.id
-            })
-            .filter((lineId) => {
-              return (
-                file_with_two_levels.lines
-                  .map((line) => {
-                    return line.id
-                  })
-                  .indexOf(lineId) === -1
-              )
-            })
-          const addedCards = result.cards.filter((card) => {
-            return templateLines.indexOf(card.lineId) !== -1
-          })
-          const templateCards = one_level_biased_top_plotline_template.templateData.cards
-          expect(templateCards.length).toEqual(addedCards.length)
-          expect(addedCards.map(withoutLineBeatOrCardIds)).toEqual(
-            expect.arrayContaining(templateCards.map(withoutLineBeatOrCardIds))
-          )
-        })
-      })
-      describe('and the template is biased to the "middle"', () => {
-        const result = applyTemplate(
-          file_with_two_levels,
-          1,
-          one_level_biased_middle_plotline_template
-        )
-        const templateBeats = tree.children(
-          one_level_biased_middle_plotline_template.templateData.beats['1'],
-          null
-        )
-        it('should have more cards', () => {
-          expect(result.cards.length).toBeGreaterThanOrEqual(file_with_two_levels.cards.length)
-        })
-        it('should have the cards it used to have', () => {
-          expect(result.cards).toEqual(expect.arrayContaining(file_with_two_levels.cards))
-        })
-        it('should have the cards in the template', () => {
-          expect(result.cards.map(withoutLineBeatOrCardIds)).toEqual(
-            expect.arrayContaining(
-              one_level_biased_middle_plotline_template.templateData.cards.map(
-                withoutLineBeatOrCardIds
-              )
-            )
-          )
-        })
-        it('should have at least as many beats as the top level of the template', () => {
-          const templateNumberOfBeats = templateBeats.length
-          const resultNumberOfBeats = beatsLength(result.beats['1'])
-          expect(resultNumberOfBeats).toBeGreaterThanOrEqual(templateNumberOfBeats)
-        })
-        it('should have the plotlines that were there originally', () => {
-          expect(result.lines).toEqual(expect.arrayContaining(file_with_two_levels.lines))
-        })
-        it('should insert cards on the plotlines from the template', () => {
-          const templateLines = result.lines
-            .map((line) => {
-              return line.id
-            })
-            .filter((lineId) => {
-              return (
-                file_with_two_levels.lines
-                  .map((line) => {
-                    return line.id
-                  })
-                  .indexOf(lineId) === -1
-              )
-            })
-          const addedCards = result.cards.filter((card) => {
-            return templateLines.indexOf(card.lineId) !== -1
-          })
-          const templateCards = one_level_biased_middle_plotline_template.templateData.cards
-          expect(templateCards.length).toEqual(addedCards.length)
-          expect(addedCards.map(withoutLineBeatOrCardIds)).toEqual(
-            expect.arrayContaining(templateCards.map(withoutLineBeatOrCardIds))
-          )
-        })
-      })
-      describe('and the template is biased to the "bottom"', () => {
-        const result = applyTemplate(
-          file_with_two_levels,
-          1,
-          one_level_biased_bottom_plotline_template
-        )
-        const templateBeats = tree.children(file_with_two_levels.templateData.beats['1'], null)
-        it('should have more cards', () => {
-          expect(result.cards.length).toBeGreaterThanOrEqual(file_with_two_levels.cards.length)
-        })
-        it('should have the cards it used to have', () => {
-          expect(result.cards).toEqual(expect.arrayContaining(file_with_two_levels.cards))
-        })
-        it('should have the cards in the template', () => {
-          expect(result.cards.map(withoutLineBeatOrCardIds)).toEqual(
-            expect.araryContaining(
-              one_level_biased_bottom_plotline_template.templateData.cards.map(
-                withoutLineBeatOrCardIds
-              )
-            )
-          )
-        })
-        it('should have the plotlines that were there originally', () => {
-          expect(result.lines).toEqual(expect.arrayContaining(file_with_two_levels.lines))
-        })
-        it('should insert cards on the plotlines from the template', () => {
-          const templateLines = result.lines
-            .map((line) => {
-              return line.id
-            })
-            .filter((lineId) => {
-              return (
-                file_with_two_levels.lines
-                  .map((line) => {
-                    return line.id
-                  })
-                  .indexOf(lineId) === -1
-              )
-            })
-          const addedCards = result.cards.filter((card) => {
-            return templateLines.indexOf(card.lineId) !== -1
-          })
-          const templateCards = one_level_biased_bottom_plotline_template.templateData.cards
-          expect(templateCards.length).toEqual(addedCards.length)
-          expect(addedCards.map(withoutLineBeatOrCardIds)).toEqual(
-            expect.arrayContaining(templateCards.map(withoutLineBeatOrCardIds))
-          )
-        })
-      })
-    })
-    describe('and a template with two levels', () => {
-      describe('and the template is biased to the "top"', () => {
-        const result = applyTemplate(
-          file_with_two_levels,
-          1,
-          two_level_biased_top_plotline_template
-        )
-        const templateTopBeats = tree.children(
-          two_level_biased_top_plotline_template.templateData.beats['1'],
-          null
-        )
-        const templateBottomBeats = templateTopBeats.flatMap((beat) => {
-          return tree.children(templateTopBeats.templateData.beats['1'], beat.id)
-        })
-        const resultTopBeats = tree.children(result.beats['1'], null)
-        const resultBottomBeats = resultTopBeats.flatMap((beat) => {
-          return tree.children(result.beats['1'], beat.id)
-        })
-        it('should have more cards', () => {
-          expect(result.cards.length).toBeGreaterThanOrEqual(file_with_two_levels.cards.length)
-        })
-        it('should have the cards it used to have', () => {
-          expect(result.cards).toEqual(expect.arrayContaining(file_with_two_levels.cards))
-        })
-        it('should have the cards in the template', () => {
-          expect(result.cards.map(withoutLineBeatOrCardIds)).toEqual(
-            expect.arrayContaining(
-              two_level_biased_top_plotline_template.templateData.cards.map(
-                withoutLineBeatOrCardIds
-              )
-            )
-          )
-        })
-        it('should have at least as many beats as the template at the top level', () => {
-          const templateNumberOfBeats = templateTopBeats.length
-          const resultNumberOfBeats = resultTopBeats.length
-          expect(resultNumberOfBeats).toBeGreaterThanOrEqual(templateNumberOfBeats)
-        })
-        it('should have at least as many beats as the template in the middle', () => {
-          const templateNumberOfBeats = templateBottomBeats.length
-          const resultNumberOfBeats = resultBottomBeats.length
-          expect(resultNumberOfBeats).toBeGreaterThanOrEqual(templateNumberOfBeats)
-        })
-        it('should have the plotlines that were there originally', () => {
-          expect(result.lines).toEqual(expect.arrayContaining(file_with_two_levels.lines))
-        })
-        it('should insert cards on the plotlines from the template', () => {
-          const templateLines = result.lines
-            .map((line) => {
-              return line.id
-            })
-            .filter((lineId) => {
-              return (
-                file_with_two_levels.lines
-                  .map((line) => {
-                    return line.id
-                  })
-                  .indexOf(lineId) === -1
-              )
-            })
-          const addedCards = result.cards.filter((card) => {
-            return templateLines.indexOf(card.lineId) !== -1
-          })
-          const templateCards = two_level_biased_bottom_plotline_template.templateData.cards
-          expect(templateCards.length).toEqual(addedCards.length)
-          expect(addedCards.map(withoutLineBeatOrCardIds)).toEqual(
-            expect.arrayContaining(templateCards.map(withoutLineBeatOrCardIds))
-          )
-        })
-      })
-      describe('and the template is biased to the middle', () => {
-        const result = applyTemplate(
-          file_with_two_levels,
-          1,
-          two_level_biased_middle_plotline_template
-        )
-        const templateTopBeats = tree.children(
-          two_level_biased_middle_plotline_template.templateData.beats['1'],
-          null
-        )
-        const templateBottomBeats = templateTopBeats.flatMap((beat) => {
-          return tree.children(
-            two_level_biased_middle_plotline_template.templateData.beats['1'],
-            beat.id
-          )
-        })
-        const resultTopBeats = tree.children(result.beats['1'], null)
-        const resultBottomBeats = resultTopBeats.flatMap((beat) => {
-          return tree.children(result.beats['1'], beat.id)
-        })
-        it('should have more cards', () => {
-          expect(result.cards.length).toBeGreaterThanOrEqual(file_with_two_levels.cards.length)
-        })
-        it('should have the cards it used to have', () => {
-          expect(result.cards).toEqual(expect.arrayContaining(file_with_two_levels.cards))
-        })
-        it('should have at least as many beats as the template at the top level', () => {
-          const templateNumberOfBeats = templateTopBeats.length
-          const resultNumberOfBeats = resultTopBeats.length
-          expect(resultNumberOfBeats).toBeGreaterThanOrEqual(templateNumberOfBeats)
-        })
-        it('should have at least as many beats as the template in the middle', () => {
-          const templateNumberOfBeats = templateBottomBeats.length
-          const resultNumberOfBeats = resultBottomBeats.length
-          expect(templateNumberOfBeats).toBeGreaterThanOrEqual(resultNumberOfBeats)
-        })
-        it('should have the plotlines that were there originally', () => {
-          expect(result.lines).toEqual(expect.arrayContaining(file_with_two_levels.lines))
-        })
-        it('should insert cards on the plotlines from the template', () => {
-          const templateLines = result.lines
-            .map((line) => {
-              return line.id
-            })
-            .filter((lineId) => {
-              return (
-                file_with_two_levels.lines
-                  .map((line) => {
-                    return line.id
-                  })
-                  .indexOf(lineId) === -1
-              )
-            })
-          const addedCards = result.cards.filter((card) => {
-            return templateLines.indexOf(card.lineId) !== -1
-          })
-          const templateCards = two_level_biased_middle_plotline_template.templateData.cards
-          expect(templateCards.length).toEqual(addedCards.length)
-          expect(addedCards.map(withoutLineBeatOrCardIds)).toEqual(
-            expect.arrayContaining(templateCards.map(withoutLineBeatOrCardIds))
-          )
-        })
-      })
-      describe('and the template is biased to the bottom', () => {
-        const result = applyTemplate(
-          file_with_two_levels,
-          1,
-          two_level_biased_bottom_plotline_template
-        )
-        const templateTopBeats = tree.children(
-          two_level_biased_bottom_plotline_template.templateData.beats['1'],
-          null
-        )
-        const templateBottomBeats = templateTopBeats.flatMap((beat) => {
-          return tree.children(two_level_biased_bottom_plotline_template.beats['1'], beat.id)
-        })
-        const resultTopBeats = tree.children(result.beats['1'], null)
-        const resultBottomBeats = resultTopBeats.flatMap((beat) => {
-          return tree.children(file_with_two_levels.beats['1'], beat.id)
-        })
-        it('should have more cards', () => {
-          expect(result.cards.length).toBeGreaterThanOrEqual(file_with_two_levels.cards.length)
-        })
-        it('should have the cards it used to have', () => {
-          expect(result.cards).toEqual(expect.arrayContaining(file_with_two_levels.cards))
-        })
-        it('should have the cards in the template', () => {
-          expect(result.cards.map(withoutLineBeatOrCardIds)).toEqual(
-            expect.arrayContaining(
-              two_level_biased_bottom_plotline_template.cards.map(withoutLineBeatOrCardIds)
-            )
-          )
-        })
-        it('should have at least as many beats as the template at the top level', () => {
-          const templateNumberOfBeats = templateTopBeats.length
-          const resultNumberOfBeats = resultTopBeats.length
-          expect(resultNumberOfBeats).toBeGreaterThanOrEqual(templateNumberOfBeats)
-        })
-        it('should have at least as many beats as the template in the middle', () => {
-          const templateNumberOfBeats = templateBottomBeats.length
-          const resultNumberOfBeats = resultBottomBeats.length
-          expect(resultNumberOfBeats).toBeGreaterThanOrEqual(templateNumberOfBeats)
-        })
-        it('should have the plotlines that were there originally', () => {
-          expect(result.lines).toEqual(
-            expect.arrayContaining(two_level_biased_bottom_plotline_template.lines)
-          )
-        })
-        it('should insert cards on the plotlines from the template', () => {
-          const templateLines = result.lines
-            .map((line) => {
-              return line.id
-            })
-            .filter((lineId) => {
-              return (
-                file_with_two_levels.lines
-                  .map((line) => {
-                    return line.id
-                  })
-                  .indexOf(lineId) === -1
-              )
-            })
-          const addedCards = result.cards.filter((card) => {
-            return templateLines.indexOf(card.lineId) !== -1
-          })
-          const templateCards = two_level_biased_bottom_plotline_template.templateData.cards
-          expect(templateCards.length).toEqual(addedCards.length)
-          expect(addedCards.map(withoutLineBeatOrCardIds)).toEqual(
-            expect.arrayContaining(templateCards.map(withoutLineBeatOrCardIds))
-          )
-        })
-      })
-    })
-    describe('and a template with three levels', () => {
-      describe('and the template is biased to the top', () => {
-        const result = applyTemplate(
-          file_with_two_levels,
-          1,
-          three_level_biased_top_plotline_template
-        )
-        const templateTopBeats = tree.children(
-          three_level_biased_top_plotline_template.templateData.beats['1'],
-          null
-        )
-        const templateMiddleBeats = templateTopBeats.flatMap((beat) => {
-          return tree.children(
-            three_level_biased_top_plotline_template.templateData.beats['1'],
-            beat.id
-          )
-        })
-        const resultTopBeats = tree.children(result.beats['1'], null)
-        const resultBottomBeats = resultTopBeats.flatMap((beat) => {
-          return tree.children(result.beats['1'], beat.id)
-        })
-        it('should have more cards', () => {
-          expect(result.cards.length).toBeGreaterThanOrEqual(file_with_two_levels.cards.length)
-        })
-        it('should have the cards it used to have', () => {
-          expect(result.cards).toEqual(expect.arrayContaining(file_with_two_levels.cards))
-        })
-        it('should have at least as many beats as the template at the top level', () => {
-          const templateNumberOfBeats = templateTopBeats.length
-          const resultNumberOfBeats = resultTopBeats.length
-          expect(resultNumberOfBeats).toBeGreaterThanOrEqual(templateNumberOfBeats)
-        })
-        it('should have at least as many beats as the template in the middle', () => {
-          const templateNumberOfBeats = templateMiddleBeats.length
-          const resultNumberOfBeats = resultBottomBeats.length
-          expect(resultNumberOfBeats).toBeGreaterThanOrEqual(templateNumberOfBeats)
-        })
-        it('should have the plotlines that were there originally', () => {
-          expect(result.lines).toEqual(
-            expect.arrayContaining(three_level_biased_top_plotline_template.lines)
-          )
-        })
-        it('should insert cards on the plotlines from the template', () => {
-          const templateLines = result.lines
-            .map((line) => {
-              return line.id
-            })
-            .filter((lineId) => {
-              return (
-                file_with_two_levels.lines
-                  .map((line) => {
-                    return line.id
-                  })
-                  .indexOf(lineId) === -1
-              )
-            })
-          const addedCards = result.cards.filter((card) => {
-            return templateLines.indexOf(card.lineId) !== -1
-          })
-          const templateCards = three_level_biased_top_plotline_template.templateData.cards
-          expect(templateCards.length).toEqual(addedCards.length)
-          expect(addedCards.map(withoutLineBeatOrCardIds)).toEqual(
-            expect.arrayContaining(templateCards.map(withoutLineBeatOrCardIds))
-          )
-        })
-      })
-      describe('and the template is biased to the middle', () => {
-        const result = applyTemplate(
-          file_with_two_levels,
-          1,
-          three_level_biased_middle_plotline_template
-        )
-        const templateTopBeats = tree.children(
-          three_level_biased_middle_plotline_template.templateData.beats['1'],
-          null
-        )
-        const templateMiddleBeats = templateTopBeats.flatMap((beat) => {
-          return tree.children(
-            three_level_biased_middle_plotline_template.templateData.beats['1'],
-            beat.id
-          )
-        })
-        const resultTopBeats = tree.children(result.beats['1'], null)
-        const resultBottomBeats = resultTopBeats.flatMap((beat) => {
-          return tree.children(result.beats['1'], beat.id)
-        })
-        it('should have more cards', () => {
-          expect(result.cards.length).toBeGreaterThanOrEqual(file_with_two_levels.cards.length)
-        })
-        it('should have the cards it used to have', () => {
-          expect(result.cards).toEqual(expect.arrayContaining(file_with_two_levels.cards))
-        })
-        it('should have at least as many beats as the template at the top level', () => {
-          const templateNumberOfBeats = templateTopBeats.length
-          const resultNumberOfBeats = resultTopBeats.length
-          expect(resultNumberOfBeats).toBeGreaterThanOrEqual(templateNumberOfBeats)
-        })
-        it('should have at least as many beats as the template in the middle', () => {
-          const templateNumberOfBeats = templateMiddleBeats.length
-          const resultNumberOfBeats = resultBottomBeats.leangth
-          expect(resultNumberOfBeats).toBeGreaterThanOrEqual(templateNumberOfBeats)
-        })
-        it('should have the plotlines that were there originally', () => {
-          expect(result.lines).toEqual(expect.arrayContaining(file_with_two_levels.lines))
-        })
-        it('should insert cards on the plotlines from the template', () => {
-          const templateLines = result.lines
-            .map((line) => {
-              return line.id
-            })
-            .filter((lineId) => {
-              return (
-                file_with_two_levels.lines
-                  .map((line) => {
-                    return line.id
-                  })
-                  .indexOf(lineId) === -1
-              )
-            })
-          const addedCards = result.cards.filter((card) => {
-            return templateLines.indexOf(card.lineId) !== -1
-          })
-          const templateCards = three_level_biased_middle_plotline_template.templateData.cards
-          expect(templateCards.length).toEqual(addedCards.length)
-          expect(addedCards.map(withoutLineBeatOrCardIds)).toEqual(
-            expect.arrayContaining(templateCards.map(withoutLineBeatOrCardIds))
-          )
-        })
-      })
-      describe('and the template is biased to the bottom', () => {
-        const result = applyTemplate(
-          file_with_two_levels,
-          1,
-          three_level_biased_bottom_plotline_template
-        )
-        const templateTopBeats = tree.children(
-          three_level_biased_middle_plotline_template.templateData.beats['1'],
-          null
-        )
-        const templateMiddleBeats = templateTopBeats.flatMap((beat) => {
-          return tree.children(
-            three_level_biased_middle_plotline_template.templateData.beats['1'],
-            beat.id
-          )
-        })
-        const templateBottomBeats = templateMiddleBeats.flatMap((beat) => {
-          return tree.children(
-            three_level_biased_middle_plotline_template.templateData.beats['1'],
-            beat.id
-          )
-        })
-        const resultTopBeats = tree.children(result.beats['1'], null)
-        const resultBottomBeats = resultTopBeats.flatMap((beat) => {
-          return tree.children(result.beats['1'], beat.id)
-        })
-        it('should have more cards', () => {
-          expect(result.cards.length).toBeGreaterThanOrEqual(file_with_two_levels.cards.length)
-        })
-        it('should have the cards it used to have', () => {
-          expect(result.cards).toEqual(expect.arrayContaining(file_with_two_levels.cards))
-        })
-        it('should have at least as many top beats as the middle level of the template', () => {
-          const templateNumberOfBeats = templateMiddleBeats.length
-          const resultNumberOfBeats = resultTopBeats.length
-          expect(resultNumberOfBeats).toBeGreaterThanOrEqual(templateNumberOfBeats)
-        })
-        it('should have at least as many bottom beats as the bottom level of the template', () => {
-          const templateNumberOfBeats = templateBottomBeats.length
-          const resultNumberOfBeats = resultBottomBeats.length
-          expect(resultNumberOfBeats).toBeGreaterThanOrEqual(templateNumberOfBeats)
-        })
-        it('should have the plotlines that were there originally', () => {
-          expect(result.lines).toEqual(expect.arrayContaining(file_with_two_levels.lines))
-        })
-        it('should insert cards on the plotlines from the template', () => {
-          const templateLines = result.lines
-            .map((line) => {
-              return line.id
-            })
-            .filter((lineId) => {
-              return (
-                file_with_two_levels.lines
-                  .map((line) => {
-                    return line.id
-                  })
-                  .indexOf(lineId) === -1
-              )
-            })
-          const addedCards = result.cards.filter((card) => {
-            return templateLines.indexOf(card.lineId) !== -1
-          })
-          const templateCards = three_level_biased_bottom_plotline_template.templateData.cards
-          expect(templateCards.length).toEqual(addedCards.length)
-          expect(addedCards.map(withoutLineBeatOrCardIds)).toEqual(
-            expect.arrayContaining(templateCards.map(withoutLineBeatOrCardIds))
-          )
-        })
-      })
-    })
-  })
-  describe('given a file with three levels', () => {
-    describe('and a template with one level', () => {
-      describe('and the template is biased to the "top"', () => {
-        const result = applyTemplate(
-          file_with_three_levels,
-          1,
-          one_level_biased_top_plotline_template
-        )
-        const templateBeats = tree.children(
-          one_level_biased_top_plotline_template.templateData.beats['1'],
-          null
-        )
-        const resultTopBeats = tree.children(result.beats['1'], null)
-        it('should have more cards', () => {
-          expect(result.cards.length).toBeGreaterThanOrEqual(file_with_three_levels.cards.length)
-        })
-        it('should have the cards it used to have', () => {
-          expect(result.cards).toEqual(expect.arrayContaining(file_with_three_levels.cards))
-        })
-        it('should have the cards in the template', () => {
-          expect(result.cards.map(withoutLineBeatOrCardIds)).toEqual(
-            expect.arrayContaining(
-              one_level_biased_top_plotline_template.templateData.cards.map(
-                withoutLineBeatOrCardIds
-              )
-            )
-          )
-        })
-        it('should have at least as many top beats as the template', () => {
-          const templateNumberOfBeats = templateBeats.length
-          const resultNumberOfBeats = resultTopBeats.length
-          expect(resultNumberOfBeats).toBeGreaterThanOrEqual(templateNumberOfBeats)
-        })
-        it('should have the plotlines that were there originally', () => {
-          expect(result.lines).toEqual(expect.arrayContaining(file_with_three_levels.lines))
-        })
-        it('should insert cards on the plotlines from the template', () => {
-          const templateLines = result.lines
-            .map((line) => {
-              return line.id
-            })
-            .filter((lineId) => {
-              return (
-                file_with_three_levels.lines
-                  .map((line) => {
-                    return line.id
-                  })
-                  .indexOf(lineId) === -1
-              )
-            })
-          const addedCards = result.cards.filter((card) => {
-            return templateLines.indexOf(card.lineId) !== -1
-          })
-          const templateCards = one_level_biased_top_plotline_template.templateData.cards
-          expect(templateCards.length).toEqual(addedCards.length)
-          expect(addedCards.map(withoutLineBeatOrCardIds)).toEqual(
-            expect.arrayContaining(templateCards.map(withoutLineBeatOrCardIds))
-          )
-        })
-      })
-      describe('and the template is biased to the "middle"', () => {
-        const result = applyTemplate(
-          file_with_three_levels,
-          1,
-          one_level_biased_middle_plotline_template
-        )
-        const templateBeats = tree.children(
-          one_level_biased_middle_plotline_template.templateData.beats['1'],
-          null
-        )
-        const resultTopBeats = tree.children(result.beats['1'], null)
-        const resultMiddleBeats = resultTopBeats.flatMap((beat) => {
-          return tree.children(result.beats['1'], beat.id)
-        })
-        it('should have more cards', () => {
-          expect(result.cards.length).toBeGreaterThanOrEqual(file_with_three_levels.cards.length)
-        })
-        it('should have the cards it used to have', () => {
-          expect(result.cards).toEqual(expect.arrayContaining(file_with_three_levels.cards))
-        })
-        it('should have the cards in the template', () => {
-          expect(result.cards.map(withoutLineBeatOrCardIds)).toEqual(
-            expect.arrayContaining(
-              one_level_biased_middle_plotline_template.templateData.cards.map(
-                withoutLineBeatOrCardIds
-              )
-            )
-          )
-        })
-        it('should have at least as many middle beats as the template', () => {
-          const templateNumberOfBeats = templateBeats.length
-          const resultNumberOfBeats = resultMiddleBeats.length
-          expect(resultNumberOfBeats).toBeGreaterThanOrEqual(templateNumberOfBeats)
-        })
-        it('should have the plotlines that were there originally', () => {
-          expect(result.lines).toEqual(expect.arrayContaining(file_with_three_levels.lines))
-        })
-        it('should insert cards on the plotlines from the template', () => {
-          const templateLines = result.lines
-            .map((line) => {
-              return line.id
-            })
-            .filter((lineId) => {
-              return (
-                file_with_three_levels.lines
-                  .map((line) => {
-                    return line.id
-                  })
-                  .indexOf(lineId) === -1
-              )
-            })
-          const addedCards = result.cards.filter((card) => {
-            return templateLines.indexOf(card.lineId) !== -1
-          })
-          const templateCards = one_level_biased_middle_plotline_template.templateData.cards
-          expect(templateCards.length).toEqual(addedCards.length)
-          expect(addedCards.map(withoutLineBeatOrCardIds)).toEqual(
-            expect.arrayContaining(templateCards.map(withoutLineBeatOrCardIds))
-          )
-        })
-      })
-      describe('and the template is biased to the "bottom"', () => {
-        const result = applyTemplate(
-          file_with_three_levels,
-          1,
-          one_level_biased_bottom_plotline_template
-        )
-        const templateBeats = tree.children(
-          one_level_biased_bottom_plotline_template.templateData.beats['1'],
-          null
-        )
-        const resultTopBeats = tree.children(result.beats['1'], null)
-        const resultMiddleBeats = resultTopBeats.flatMap((beat) => {
-          return tree.children(result.beats['1'], beat.id)
-        })
-        const resultBottomBeats = resultMiddleBeats.flatMap((beat) => {
-          return tree.children(result.beats['1'], beat.id)
-        })
-        it('should have more cards', () => {
-          expect(result.cards.length).toBeGreaterThanOrEqual(file_with_three_levels.cards.length)
-        })
-        it('should have the cards it used to have', () => {
-          expect(result.cards).toEqual(expect.arrayContaining(file_with_three_levels.cards))
-        })
-        it('should have the cards in the template', () => {
-          expect(result.cards.map(withoutLineBeatOrCardIds)).toEqual(
-            expect.arrayContaining(
-              one_level_biased_bottom_plotline_template.templateData.cards.map(
-                withoutLineBeatOrCardIds
-              )
-            )
-          )
-        })
-        it('should have at least as many bottom beats as the template', () => {
-          const templateNumberOfBeats = templateBeats.length
-          const resultNumberOfBeats = resultBottomBeats.length
-          expect(resultNumberOfBeats).toBeGreaterThanOrEqual(templateNumberOfBeats)
-        })
-        it('should have the plotlines that were there originally', () => {
-          expect(result.lines).toEqual(expect.arrayContaining(file_with_three_levels.lines))
-        })
-        it('should insert cards on the plotlines from the template', () => {
-          const templateLines = result.lines
-            .map((line) => {
-              return line.id
-            })
-            .filter((lineId) => {
-              return (
-                file_with_three_levels.lines
-                  .map((line) => {
-                    return line.id
-                  })
-                  .indexOf(lineId) === -1
-              )
-            })
-          const addedCards = result.cards.filter((card) => {
-            return templateLines.indexOf(card.lineId) !== -1
-          })
-          const templateCards = one_level_biased_bottom_plotline_template.templateData.cards
-          expect(templateCards.length).toEqual(addedCards.length)
-          expect(addedCards.map(withoutLineBeatOrCardIds)).toEqual(
-            expect.arrayContaining(templateCards.map(withoutLineBeatOrCardIds))
-          )
-        })
-      })
-    })
-    describe('and a template with two levels', () => {
-      describe('and the template is biased to the "top"', () => {
-        const result = applyTemplate(
-          file_with_three_levels,
-          1,
-          two_level_biased_top_plotline_template
-        )
-        const templateTopBeats = tree.children(
-          two_level_biased_top_plotline_template.templateData.beats['1'],
-          null
-        )
-        const templateBottomBeats = templateTopBeats.flatMap((beat) => {
-          return tree.children(
-            two_level_biased_top_plotline_template.templateData.beats['1'],
-            beat.id
-          )
-        })
-        const resultTopBeats = tree.children(result.beats['1'], null)
-        const resultMiddleBeats = resultTopBeats.flatMap((beat) => {
-          return tree.children(result.beats['1'], beat.id)
-        })
-        it('should have more cards', () => {
-          expect(result.cards.length).toBeGreaterThanOrEqual(file_with_three_levels.cards.length)
-        })
-        it('should have the cards in the template', () => {
-          expect(result.cards.map(withoutLineBeatOrCardIds)).toEqual(
-            expect.arrayContaining(
-              two_level_biased_top_plotline_template.templateData.cards.map(
-                withoutLineBeatOrCardIds
-              )
-            )
-          )
-        })
-        it('should have at least as many top beats as the template top beats', () => {
-          const templateNumberOfBeats = templateTopBeats.length
-          const resultNumberOfBeats = resultTopBeats.length
-          expect(resultNumberOfBeats).toBeGreaterThanOrEqual(templateNumberOfBeats)
-        })
-        it('should have at least as many middle beats as the template bottom beats', () => {
-          const templateNumberOfBeats = templateBottomBeats.length
-          const resultNumberOfBeats = resultMiddleBeats.length
-          expect(resultNumberOfBeats).toBeGreaterThanOrEqual(templateNumberOfBeats)
-        })
-        it('should have the plotlines that were there originally', () => {
-          expect(result.lines).toEqual(expect.arrayContaining(file_with_three_levels.lines))
-        })
-        it('should insert cards on the plotlines from the template', () => {
-          const templateLines = result.lines
-            .map((line) => {
-              return line.id
-            })
-            .filter((lineId) => {
-              return (
-                file_with_three_levels.lines
-                  .map((line) => {
-                    return line.id
-                  })
-                  .indexOf(lineId) === -1
-              )
-            })
-          const addedCards = result.cards.filter((card) => {
-            return templateLines.indexOf(card.lineId) !== -1
-          })
-          const templateCards = two_level_biased_top_plotline_template.templateData.cards
-          expect(templateCards.length).toEqual(addedCards.length)
-          expect(addedCards.map(withoutLineBeatOrCardIds)).toEqual(
-            expect.arrayContaining(templateCards.map(withoutLineBeatOrCardIds))
-          )
-        })
-      })
-      describe('and the template is biased to the middle', () => {
-        const result = applyTemplate(
-          file_with_three_levels,
-          1,
-          two_level_biased_middle_plotline_template
-        )
-        const templateTopBeats = tree.children(
-          two_level_biased_middle_plotline_template.templateData.beats['1'],
-          null
-        )
-        const templateBottomBeats = templateTopBeats.flatMap((beat) => {
-          return tree.children(
-            two_level_biased_middle_plotline_template.templateData.beats['1'],
-            beat.id
-          )
-        })
-        const resultTopBeats = tree.children(result.beats['1'], null)
-        const resultMiddleBeats = resultTopBeats.flatMap((beat) => {
-          return tree.children(result.beats['1'], beat.id)
-        })
-        const resultBottomBeats = resultMiddleBeats.flatMap((beat) => {
-          return tree.children(result.beats['1'], beat.id)
-        })
-        it('should have more cards', () => {
-          expect(result.cards.length).toBeGreaterThanOrEqual(file_with_three_levels.cards.length)
-        })
-        it('should have the cards in the template', () => {
-          expect(result.cards.map(withoutLineBeatOrCardIds)).toEqual(
-            expect.arrayContaining(
-              two_level_biased_middle_plotline_template.templateData.cards.map(
-                withoutLineBeatOrCardIds
-              )
-            )
-          )
-        })
-        it('should have at least as many middle beats as the template top beats', () => {
-          const templateNumberOfBeats = templateTopBeats.length
-          const resultNumberOfBeats = resultMiddleBeats.length
-          expect(resultNumberOfBeats).toBeGreaterThanOrEqual(templateNumberOfBeats)
-        })
-        it('should have at least as many bottom beats as the template bottom beats', () => {
-          const templateNumberOfBeats = templateBottomBeats.length
-          const resultNumberOfBeats = resultBottomBeats.length
-          expect(resultNumberOfBeats).toBeGreaterThanOrEqual(templateNumberOfBeats)
-        })
-        it('should have the plotlines that were there originally', () => {
-          expect(result.lines).toEqual(expect.arrayContaining(file_with_three_levels))
-        })
-        it('should insert cards on the plotlines from the template', () => {
-          const templateLines = result.lines
-            .map((line) => {
-              return line.id
-            })
-            .filter((lineId) => {
-              return (
-                file_with_three_levels.lines
-                  .map((line) => {
-                    return line.id
-                  })
-                  .indexOf(lineId) === -1
-              )
-            })
-          const addedCards = result.cards.filter((card) => {
-            return templateLines.indexOf(card.lineId) !== -1
-          })
-          const templateCards = two_level_biased_middle_plotline_template.templateData.cards
-          expect(templateCards.length).toEqual(addedCards.length)
-          expect(addedCards.map(withoutLineBeatOrCardIds)).toEqual(
-            expect.arrayContaining(templateCards.map(withoutLineBeatOrCardIds))
-          )
-        })
-      })
-      describe('and the template is biased to the bottom', () => {
-        const result = applyTemplate(
-          file_with_three_levels,
-          1,
-          two_level_biased_bottom_plotline_template
-        )
-        const templateTopBeats = tree.children(
-          two_level_biased_bottom_plotline_template.templateData.beats['1'],
-          null
-        )
-        const templateBottomBeats = templateTopBeats.flatMap((beat) => {
-          return tree.children(
-            two_level_biased_bottom_plotline_template.templateData.beats['1'],
-            beat.id
-          )
-        })
-        const resultTopBeats = tree.children(result.beats['1'], null)
-        const resultMiddleBeats = resultTopBeats.flatMap((beat) => {
-          return tree.children(result.beats['1'], beat.id)
-        })
-        const resultBottomBeats = resultMiddleBeats.flatMap((beat) => {
-          return tree.children(result.beats['1'], beat.id)
-        })
-        it('should have more cards', () => {
-          expect(result.cards.length).toBeGreaterThanOrEqual(file_with_three_levels.cards.length)
-        })
-        it('should have the cards in the template', () => {
-          expect(result.card.map(withoutLineBeatOrCardIds)).toEqual(
-            expect.arrayContaining(
-              two_level_biased_bottom_plotline_template.templateData.cards.map(
-                withoutLineBeatOrCardIds
-              )
-            )
-          )
-        })
-        it('should have at least as many middle beats as the template top beats', () => {
-          const templateNumberOfBeats = templateTopBeats.length
-          const resultNumberOfBeats = resultMiddleBeats.length
-          expect(resultNumberOfBeats).toBeGreaterThanOrEqual(templateNumberOfBeats)
-        })
-        it('should have at least as many bottom beats as the template bottom beats', () => {
-          const templateNumberOfBeats = templateBottomBeats.length
-          const resultNumberOfBeats = resultBottomBeats.length
-          expect(resultNumberOfBeats).toBeGreaterThanOrEqual(templateNumberOfBeats)
-        })
-        it('should have the plotlines that were there originally', () => {
-          expect(result.lines).toEqual(expect.arrayContaining(file_with_three_levels.lines))
-        })
-        it('should insert cards on the plotlines from the template', () => {
-          const templateLines = result.lines
-            .map((line) => {
-              return line.id
-            })
-            .filter((lineId) => {
-              return (
-                file_with_three_levels.lines
-                  .map((line) => {
-                    return line.id
-                  })
-                  .indexOf(lineId) === -1
-              )
-            })
-          const addedCards = result.cards.filter((card) => {
-            return templateLines.indexOf(card.lineId) !== -1
-          })
-          const templateCards = two_level_biased_bottom_plotline_template.templateData.cards
-          expect(templateCards.length).toEqual(addedCards.length)
-          expect(addedCards.map(withoutLineBeatOrCardIds)).toEqual(
-            expect.arrayContaining(templateCards.map(withoutLineBeatOrCardIds))
-          )
-        })
-      })
-    })
-    describe('and a template with three levels', () => {
-      describe('and the template is biased to the top', () => {
-        const result = applyTemplate(
-          file_with_three_levels,
-          1,
-          three_level_biased_top_plotline_template
-        )
-        const templateTopBeats = tree.children(
-          three_level_biased_top_plotline_template.templateData.beats['1'],
-          null
-        )
-        const templateMiddleBeats = templateTopBeats.flatMap((beat) => {
-          return tree.children(
-            three_level_biased_top_plotline_template.templateData.beats['1'],
-            beat.id
-          )
-        })
-        const templateBottomBeats = templateMiddleBeats.flatMap((beat) => {
-          return tree.children(
-            three_level_biased_top_plotline_template.templateData.beats['1'],
-            beat.id
-          )
-        })
-        const resultTopBeats = tree.children(result.beats['1'], null)
-        const resultMiddleBeats = resultTopBeats.flatMap((beat) => {
-          return tree.children(result.beats['1'], beat.id)
-        })
-        const resultBottomBeats = resultMiddleBeats.flatMap((beat) => {
-          return tree.children(result.beats['1'], beat.id)
-        })
-        it('should have more cards', () => {
-          expect(result.cards.length).toBeGreaterThanOrEqual(file_with_three_levels.cards.length)
-        })
-        it('should have the cards in the template', () => {
-          expect(result.card.map(withoutLineBeatOrCardIds)).toEqual(
-            expect.arrayContaining(
-              three_level_biased_top_plotline_template.templateData.cards.map(
-                withoutLineBeatOrCardIds
-              )
-            )
-          )
-        })
-        it('should have at least as many top beats as the template top beats', () => {
-          const templateNumberOfBeats = templateTopBeats.length
-          const resultNumberOfBeats = resultTopBeats.length
-          expect(resultNumberOfBeats).toBeGreaterThanOrEqual(templateNumberOfBeats)
-        })
-        it('should have at least as many middle beats as the template middle beats', () => {
-          const templateNumberOfBeats = templateMiddleBeats.length
-          const resultNumberOfBeats = resultMiddleBeats.length
-          expect(resultNumberOfBeats).toBeGreaterThanOrEqual(templateNumberOfBeats)
-        })
-        it('should have at least as many bottom beats as the template bottom beats', () => {
-          const templateNumberOfBeats = templateBottomBeats.length
-          const resultNumberOfBeats = resultBottomBeats.length
-          expect(resultNumberOfBeats).toBeGreaterThanOrEqual(templateNumberOfBeats)
-        })
-        it('should have the plotlines that were there originally', () => {
-          expect(result.lines).toEqual(expect.arrayContaining(file_with_three_levels.lines))
-        })
-        it('should insert cards on the plotlines from the template', () => {
-          const templateLines = result.lines
-            .map((line) => {
-              return line.id
-            })
-            .filter((lineId) => {
-              return (
-                file_with_three_levels.lines
-                  .map((line) => {
-                    return line.id
-                  })
-                  .indexOf(lineId) === -1
-              )
-            })
-          const addedCards = result.cards.filter((card) => {
-            return templateLines.indexOf(card.lineId) !== -1
-          })
-          const templateCards = three_level_biased_top_plotline_template.templateData.cards
-          expect(templateCards.length).toEqual(addedCards.length)
-          expect(addedCards.map(withoutLineBeatOrCardIds)).toEqual(
-            expect.arrayContaining(templateCards.map(withoutLineBeatOrCardIds))
-          )
-        })
-      })
-      describe('and the template is biased to the middle', () => {
-        const result = applyTemplate(
-          file_with_three_levels,
-          1,
-          three_level_biased_middle_plotline_template
-        )
-        const templateTopBeats = tree.children(
-          three_level_biased_middle_plotline_template.templateData.beats['1'],
-          null
-        )
-        const templateMiddleBeats = templateTopBeats.flatMap((beat) => {
-          return tree.children(
-            three_level_biased_middle_plotline_template.templateData.beats['1'],
-            beat.id
-          )
-        })
-        const templateBottomBeats = templateMiddleBeats.flatMap((beat) => {
-          return tree.children(
-            three_level_biased_middle_plotline_template.templateData.beats['1'],
-            beat.id
-          )
-        })
-        const resultTopBeats = tree.children(result.beats['1'], null)
-        const resultMiddleBeats = resultTopBeats.flatMap((beat) => {
-          return tree.children(result.beats['1'], beat.id)
-        })
-        const resultBottomBeats = resultMiddleBeats.flatMap((beat) => {
-          return tree.children(result.beats['1'], beat.id)
-        })
-        it('should have more cards', () => {
-          expect(result.cards.length).toBeGreaterThanOrEqual(file_with_three_levels.cards.length)
-        })
-        it('should have the cards in the template', () => {
-          expect(result.card.map(withoutLineBeatOrCardIds)).toEqual(
-            expect.arrayContaining(
-              three_level_biased_middle_plotline_template.templateData.cards.map(
-                withoutLineBeatOrCardIds
-              )
-            )
-          )
-        })
-        it('should have at least as many top beats as the template top beats', () => {
-          const templateNumberOfBeats = templateTopBeats.length
-          const resultNumberOfBeats = resultTopBeats.length
-          expect(resultNumberOfBeats).toBeGreaterThanOrEqual(templateNumberOfBeats)
-        })
-        it('should have at least as many middle beats as the template middle beats', () => {
-          const templateNumberOfBeats = templateMiddleBeats.length
-          const resultNumberOfBeats = resultMiddleBeats.length
-          expect(resultNumberOfBeats).toBeGreaterThanOrEqual(templateNumberOfBeats)
-        })
-        it('should have at least as many bottom beats as the template bottom beats', () => {
-          const templateNumberOfBeats = templateBottomBeats.length
-          const resultNumberOfBeats = resultBottomBeats.length
-          expect(resultNumberOfBeats).toBeGreaterThanOrEqual(templateNumberOfBeats)
-        })
-        it('should have the plotlines that were there originally', () => {
-          expect(result.lines).toEqual(expect.arrayContaining(file_with_three_levels.lines))
-        })
-        it('should insert cards on the plotlines from the template', () => {
-          const templateLines = result.lines
-            .map((line) => {
-              return line.id
-            })
-            .filter((lineId) => {
-              return (
-                file_with_three_levels.lines
-                  .map((line) => {
-                    return line.id
-                  })
-                  .indexOf(lineId) === -1
-              )
-            })
-          const addedCards = result.cards.filter((card) => {
-            return templateLines.indexOf(card.lineId) !== -1
-          })
-          const templateCards = three_level_biased_middle_plotline_template.templateData.cards
-          expect(templateCards.length).toEqual(addedCards.length)
-          expect(addedCards.map(withoutLineBeatOrCardIds)).toEqual(
-            expect.arrayContaining(templateCards.map(withoutLineBeatOrCardIds))
-          )
-        })
-      })
-      describe('and the template is biased to the bottom', () => {
-        const result = applyTemplate(
-          file_with_three_levels,
-          1,
-          three_level_biased_bottom_plotline_template
-        )
-        const templateTopBeats = tree.children(
-          three_level_biased_bottom_plotline_template.templateData.beats['1'],
-          null
-        )
-        const templateMiddleBeats = templateTopBeats.flatMap((beat) => {
-          return tree.children(
-            three_level_biased_bottom_plotline_template.templateData.beats['1'],
-            beat.id
-          )
-        })
-        const templateBottomBeats = templateMiddleBeats.flatMap((beat) => {
-          return tree.children(
-            three_level_biased_bottom_plotline_template.templateData.beats['1'],
-            beat.id
-          )
-        })
-        const resultTopBeats = tree.children(result.beats['1'], null)
-        const resultMiddleBeats = resultTopBeats.flatMap((beat) => {
-          return tree.children(result.beats['1'], beat.id)
-        })
-        const resultBottomBeats = resultMiddleBeats.flatMap((beat) => {
-          return tree.children(result.beats['1'], beat.id)
-        })
-        it('should have more cards', () => {
-          expect(result.cards.length).toBeGreaterThanOrEqual(file_with_three_levels.cards.length)
-        })
-        it('should have the cards in the template', () => {
-          expect(result.card.map(withoutLineBeatOrCardIds)).toEqual(
-            expect.arrayContaining(
-              three_level_biased_bottom_plotline_template.templateData.cards.map(
-                withoutLineBeatOrCardIds
-              )
-            )
-          )
-        })
-        it('should have at least as many top beats as the template top beats', () => {
-          const templateNumberOfBeats = templateTopBeats.length
-          const resultNumberOfBeats = resultTopBeats.length
-          expect(resultNumberOfBeats).toBeGreaterThanOrEqual(templateNumberOfBeats)
-        })
-        it('should have at least as many middle beats as the template middle beats', () => {
-          const templateNumberOfBeats = templateMiddleBeats.length
-          const resultNumberOfBeats = resultMiddleBeats.length
-          expect(resultNumberOfBeats).toBeGreaterThanOrEqual(templateNumberOfBeats)
-        })
-        it('should have at least as many bottom beats as the template bottom beats', () => {
-          const templateNumberOfBeats = templateBottomBeats.length
-          const resultNumberOfBeats = resultBottomBeats.length
-          expect(resultNumberOfBeats).toBeGreaterThanOrEqual(templateNumberOfBeats)
-        })
-        it('should have the plotlines that were there originally', () => {
-          expect(result.lines).toEqual(expect.arrayContaining(file_with_three_levels.lines))
-        })
-        it('should insert cards on the plotlines from the template', () => {
-          const templateLines = result.lines
-            .map((line) => {
-              return line.id
-            })
-            .filter((lineId) => {
-              return (
-                file_with_three_levels.lines
-                  .map((line) => {
-                    return line.id
-                  })
-                  .indexOf(lineId) === -1
-              )
-            })
-          const addedCards = result.cards.filter((card) => {
-            return templateLines.indexOf(card.lineId) !== -1
-          })
-          const templateCards = three_level_biased_bottom_plotline_template.templateData.cards
-          expect(templateCards.length).toEqual(addedCards.length)
-          expect(addedCards.map(withoutLineBeatOrCardIds)).toEqual(
-            expect.arrayContaining(templateCards.map(withoutLineBeatOrCardIds))
-          )
-        })
       })
     })
   })
