+++ lib/pltr/v2/helpers/beats.js
@@ -4,8 +4,6 @@ import { beat } from '../store/initialState'
 import { isSeries as isSeriesString } from './books'
 import * as tree from '../reducers/tree'
 import { nextLevelName } from './hierarchy'
-import root from '../reducers/root'
-import { addLinesFromTemplate } from '../actions/lines'
 
 export function beatOneIsPrologue(sortedBookBeats) {
   if (!sortedBookBeats.length) return false
@@ -266,12 +264,4 @@ export const numberOfPriorChildrenAtSameDepth = (beatTree, beats, beatId) => {
   }
   if (!found) return null
   return 1 + priorChildren
-}
-
-export const applyTemplate = (fileState, bookId, template) => {
-  const templateLines = template.templateData.lines
-  const rootReducer = root({})
-  const addLinesAction = addLinesFromTemplate(template.templateData)
-  const withNewLines = rootReducer(fileState, addLinesAction)
-  return withNewLines
 }
