+++ package.json
@@ -31,7 +31,7 @@
     "extract-strings": "lib/plottr_locales/lib/format-message-cli/format-message extract -o lib/plottr_locales/src/en-extracted.json \"bin/**/*.js\" \"src/**/*.js\" \"shared/**/*.js\" \"lib/plottr_components/src/**/*.js\" \"lib/pltr/v2/**/*.js\" && npm --prefix lib/plottr_locales run sync && npm run generate-flipped",
     "generate-flipped": "node ./lib/plottr_locales/src/build-flipped-locale.js",
     "test": "jest \"(src.*)|(lib.*)|(main.*)\" && jest --runInBand --config=jest.integration.config.js \"(src.*)|(lib.*)|(main.*)\"",
-    "test:watch": "jest lib/pltr/v2/helpers/__tests__/beats.test.js --watch",
+    "test:watch": "jest lib/pltr/v2/helpers/__tests__/templates.test.js --watch",
     "test:debug": "node --inspect node_modules/.bin/jest --runInBand lib/pltr/v2/reducers/__tests__/ui.test.js",
     "test:integration": "jest test/integration",
     "lint-file": "eslint",
